# ai_tnhn/ai-api-tnhn/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Organization:
    properties:
      address:
        description: Địa chỉ
        example: 123 Lê Duẩn, Hà Nội
        type: string
      code:
        description: Mã định danh (unique)
        example: XN1
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        description: Mô tả
        example: Quản lý khu vực Hoàn Kiếm
        type: string
      drive_folder_id:
        description: Thư mục Google Drive của xí nghiệp
        example: 1abc2def3ghi
        type: string
      email:
        example: xn1@hsdc.com.vn
        type: string
      id:
        type: string
      name:
        example: Xí nghiệp Thoát nước số 1
        type: string
      order:
        description: |-
          RainStationIDs  []string `bson:"rain_station_ids" json:"rain_station_ids"`
          LakeStationIDs  []string `bson:"lake_station_ids" json:"lake_station_ids"`
          RiverStationIDs           []string `bson:"river_station_ids" json:"river_station_ids"`
          InundationIDs             []string `bson:"inundation_ids" json:"inundation_ids"`
          EmergencyConstructionIDs  []string `bson:"emergency_construction_ids" json:"emergency_construction_ids"`
          PumpingStationIDs         []string `bson:"pumping_station_ids" json:"pumping_station_ids"`
        type: string
      phone_number:
        example: "0241234567"
        type: string
      representative:
        example: Nguyễn Văn A
        type: string
      status:
        description: Trạng thái hoạt động
        example: true
        type: boolean
      updated_at:
        type: integer
    type: object
  models.Permission:
    properties:
      code:
        example: PERMISSION_ADMIN
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        example: Cho phép xem tất cả dữ liệu
        type: string
      group:
        example: Hệ thống
        type: string
      id:
        type: string
      title:
        example: Quyền quản trị
        type: string
      type:
        example: view
        type: string
      updated_at:
        type: integer
    type: object
...
  models.Role:
    properties:
      code:
        example: admin
        type: string
      created_at:
        type: integer
      deleted_at:
        type: integer
      description:
        example: Toàn quyền hệ thống
        type: string
      id:
        type: string
      is_company:
        example: true
        type: boolean
      is_employee:
        example: false
        type: boolean
      level:
        example: 1
        type: integer
      name:
        example: Quản trị viên
        type: string
      updated_at:
        type: integer
    type: object
...
  web.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: error message
        type: string
      status:
        example: error
        type: string
    type: object
  web.Response:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      status:
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Máy chủ API cho Hệ thống Giám sát Ngập lụt Hà Nội.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: API Hệ thống Giám sát Ngập lụt Hà Nội
  version: "1.0"
paths:
...
  /admin/organizations:
    get:
      description: Truy xuất danh sách các đơn vị với tùy chọn lọc
      parameters:
      - description: Lọc theo ID
        in: query
        name: id
        type: string
      - description: Lọc theo Tên
        in: query
        name: name
        type: string
      - description: Số trang
        in: query
        name: page
        type: integer
      - description: Số bản ghi mỗi trang
        in: query
        name: size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  properties:
                    data:
                      items:
                        $ref: '#/definitions/models.Organization'
                      type: array
                    total:
                      type: integer
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách đơn vị
      tags:
      - Đơn vị
    post:
      consumes:
      - application/json
      description: Tạo mới một đơn vị trong hệ thống (Chủ yếu dành cho Super Admin)
      parameters:
      - description: Dữ liệu đơn vị
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Organization'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo mới một đơn vị/xí nghiệp
      tags:
      - Đơn vị
  /admin/organizations/{id}:
    delete:
      description: Loại bỏ một đơn vị khỏi hệ thống theo ID
      parameters:
      - description: ID của đơn vị
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa đơn vị
      tags:
      - Đơn vị
    get:
      description: Truy xuất thông tin chi tiết của một đơn vị cụ thể
      parameters:
      - description: ID của đơn vị
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Organization'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy thông tin đơn vị theo ID
      tags:
      - Đơn vị
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin chi tiết của một đơn vị hiện có
      parameters:
      - description: ID của đơn vị
        in: path
        name: id
        required: true
        type: string
      - description: Dữ liệu đơn vị cập nhật
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/models.Organization'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Organization'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật thông tin đơn vị
      tags:
      - Đơn vị
  /admin/organizations/selection:
    get:
      description: Truy xuất danh sách đơn vị phục vụ hiển thị trong các ô chọn (dropdown)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  properties:
                    primary:
                      items:
                        $ref: '#/definitions/models.Organization'
                      type: array
                    shared:
                      items:
                        $ref: '#/definitions/models.Organization'
                      type: array
                  type: object
              type: object
      security:
      - BearerAuth: []
      summary: Lấy danh sách đơn vị để lựa chọn
      tags:
      - Đơn vị
...
  /permissions/matrix:
    get:
      description: Truy xuất danh sách vai trò và quyền hạn tương ứng để quản trị
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  properties:
                    permissions:
                      items:
                        $ref: '#/definitions/models.Permission'
                      type: array
                    roles:
                      items:
                        $ref: '#/definitions/models.Role'
                      type: array
                  type: object
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy ma trận quyền hạn
      tags:
      - Quyền hạn
    post:
      consumes:
      - application/json
      description: Cập nhật các quyền hạn được gán cho một vai trò cụ thể
      parameters:
      - description: Dữ liệu vai trò và danh sách mã quyền hạn
        in: body
        name: request
        required: true
        schema:
          properties:
            permissions:
              items:
                type: string
              type: array
            role:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật ma trận quyền hạn
      tags:
      - Quyền hạn
  /permissions/my:
    get:
      description: Truy xuất danh sách quyền được gán cho vai trò của người dùng hiện
        tại
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Lấy quyền hạn của tôi
      tags:
      - Quyền hạn
...
  /roles:
    get:
      description: Truy xuất danh sách các vai trò có sẵn để gán cho người dùng
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Role'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Danh sách vai trò
      tags:
      - Vai trò
    post:
      consumes:
      - application/json
      description: Tạo mới một vai trò người dùng trong hệ thống
      parameters:
      - description: Dữ liệu vai trò
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Tạo mới một vai trò
      tags:
      - Vai trò
  /roles/{id}:
    delete:
      description: Loại bỏ một vai trò khỏi hệ thống theo ID
      parameters:
      - description: ID vai trò
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Xóa vai trò
      tags:
      - Vai trò
    put:
      consumes:
      - application/json
      description: Cập nhật thông tin chi tiết của một vai trò hiện có
      parameters:
      - description: ID vai trò
        in: path
        name: id
        required: true
        type: string
      - description: Dữ liệu vai trò cập nhật
        in: body
        name: role
        required: true
        schema:
          $ref: '#/definitions/models.Role'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/web.Response'
            - properties:
                data:
                  $ref: '#/definitions/models.Role'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/web.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Cập nhật vai trò
      tags:
      - Vai trò
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"